requirements.txt
arcflash==0.1.3
asgiref==3.7.2
contourpy==1.2.0
cycler==0.12.1
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
electricpy==0.3.0
fonttools==4.49.0
jsonfield==3.1.0
kiwisolver==1.4.5
matplotlib==3.8.3
mpmath==1.3.0
numdifftools==0.9.41
numpy==1.26.4
packaging==23.2
pillow==10.2.0
protobuf==4.25.3
pyparsing==3.1.1
pyserial==3.5
python-dateutil==2.8.2
pytz==2024.1
scipy==1.12.0
six==1.16.0
sqlparse==0.4.4
sympy==1.12
typing_extensions==4.9.0
